+++ spec/requests/as_content_owner/channel_spec.rb
@@ -347,6 +347,15 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'views can be grouped by referrer' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :referrer' do
+          views = channel.views range.merge by: :referrer
+          expect(views.keys).to all(be_a String)
+        end
+      end
+
       describe 'views can be grouped by video' do
         let(:range) { {since: 4.days.ago, until: 3.days.ago} }
 
@@ -1020,6 +1029,15 @@ describe Yt::Channel, :partner do
 
         specify 'with the :by option set to :search_term' do
           estimated_minutes_watched = channel.estimated_minutes_watched range.merge by: :search_term
+          expect(estimated_minutes_watched.keys).to all(be_a String)
+        end
+      end
+
+      describe 'estimated minutes watched can be grouped by referrer' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :referrer' do
+          estimated_minutes_watched = channel.estimated_minutes_watched range.merge by: :referrer
           expect(estimated_minutes_watched.keys).to all(be_a String)
         end
       end